indexer: redeem-fees: Find wallet to redeem fees into if one exists
