+++ src/indexer/mod.rs
@@ -1,0 +1,5 @@
+//! The indexer handles the indexing and redemption of fee notes
+
+pub mod index_fees;
+pub mod queries;
+pub mod redeem_fees;
